requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework>=3.14,<4.0
psycopg2-binary==2.9.9
djangorestframework_simplejwt==5.3.1
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.5.1
tzdata==2024.1
